requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
entrypoints==0.3
flake8==3.7.8
guppy3==3.0.8
idna==2.8
importlib-metadata==0.19
isort==4.3.21
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
phue==1.1
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pytest==5.1.2
pytest-cov==2.7.1
pytest-env==0.6.2
pytest-mock==1.10.4
python-env==1.0.0
requests==2.22.0
scapy==2.4.3
schedule==0.6.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.6.0
